requirements.txt
python-telegram-bot[job-queue]==21.1.1
openai>=1.20.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
